requirements.txt
beautifulsoup4
httpx
numpy
opencv-python-headless
Pillow
playwright
python-docx
scikit-image
streamlit
